backend/api_gateway/requirements.txt
django==5.1.6
djangorestframework==3.15.1
drf-spectacular==0.27.2
requests==2.32.3
python-dotenv==1.0.1
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
redis==5.1.0
django-environ
django-redis
dj-database-url
psycopg2-binary
whitenoise
requests